requirements.txt
# 本地API服務依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3

# 資料庫連接依賴
psycopg2-binary==2.9.9  # PostgreSQL連接器
sqlalchemy==2.0.23      # ORM框架

# Streamlit應用依賴
streamlit==1.28.1
requests==2.31.0
reportlab==4.0.7

# 環境變數管理
python-dotenv==1.0.0

# 系統依賴
psutil>=5.8.0
python-multipart==0.0.6

# 注意：sqlite3 是 Python 內建模組，不需要單獨安裝